pyproject.toml
[tool.poetry]
name = "quickstart"
version = "0.1.0"
description = "A quickstart for the OLAS AI agents"
authors = ["David Vilela <dvilelaf@gmail.com>", "Viraj Patel <vptl185@gmail.com>", "Ojuswi Rastogi <ojuswimail@gmail.com>"]
readme = "README.md"
package-mode = false
packages = []
include = []

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
olas-operate-middleware = "0.15.1"
tqdm = "^4.67.1"

[tool.poetry.group.dev.dependencies]
pexpect = "^4.9.0"

[tool.pytest.ini_options]
markers = [
	"e2e: end-to-end tests that may use real network/external systems",
	"integration: integration tests that may use networked dependencies",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
